% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dp}
\alias{dp}
\title{Get distinct values and pull from data frame}
\usage{
dp(.data, var)
}
\arguments{
\item{.data}{a table of data}

\item{var}{variable name}
}
\value{
a vector
}
\description{
Get distinct values and pull from data frame
}
\examples{
dp(iris, Species)
}
