% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_split.R
\name{filter_split}
\alias{filter_split}
\title{Split a data frame on multiple conditions}
\usage{
filter_split(.x, ...)
}
\arguments{
\item{.x}{a table of data}

\item{...}{conditions}
}
\value{
a list of data frames
}
\description{
Split a data frame on multiple conditions
}
\examples{
filter_split(iris, Species == "versicolor", Species == "setosa")
}
