% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lu}
\alias{lu}
\title{Length of unique values}
\usage{
lu(x)
}
\arguments{
\item{x}{a vector}
}
\value{
numeric, length of unique values
}
\description{
Length of unique values
}
\examples{
lu(letters[1:10])
}
