% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternize.R
\name{patternize}
\alias{patternize}
\title{Reformat a string by pattern}
\usage{
patternize(x, pattern, pattern_replace = "X")
}
\arguments{
\item{x}{a string}

\item{pattern}{a string pattern}

\item{pattern_replace}{the replacement character in the pattern}
}
\value{
a string
}
\description{
Reformat a string by pattern
}
\examples{
patternize(7072999216, "(XXX) XXX-XXXX")
}
