% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{si}
\alias{si}
\title{Sum x in y}
\usage{
si(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
numeric, sum of x in y
}
\description{
Sum x in y
}
\examples{
si(c("A", "B", "C"), LETTERS)
}
