% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sni}
\alias{sni}
\title{Sum x not in y}
\usage{
sni(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
numeric, sum of x not in y
}
\description{
Sum x not in y
}
\examples{
si(c("A", "B", "C"), LETTERS)
}
